requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
